moved log out to settings and added code to stop any current music
